requirements.txt
# PyTorch Lightning and related
lightning>=2.0.0
torchmetrics>=0.11.0

# Rich console output
rich>=13.0.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0

# Visualization
matplotlib>=3.5.0
tensorboard>=2.11.0

# Data processing
pandas>=1.3.0
pillow>=9.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# For downloading pretrained model
gdown>=4.6.0

# Additional utilities
hydra-core>=1.3.0
omegaconf>=2.3.0